src/api/requirements.txt
fastapi[standard]
azure-ai-projects
azure-ai-agents
azure-identity
dotenv
aiohttp
jsonref